+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+2.2 - Oct 19, 2019
+------------------
+
+* Add message authenticator support (attribute 80)
+
+* Add support for multiple values of the same attribute (#95)
+
 * Add experimental async client and server implementation for python >=3.5.
 
 * Add IPv6 bind support for client and server.